Restock item implementation successfully done

+++ src/Pages/ManageItems/ManageItems.js
@@ -10,15 +10,18 @@ const ManageItems = () => {
         navigate('/addItem')
     }
     const handleDelete = id => {
-        const url = `http://localhost:5000/inventory/${id}`
-        fetch(url, {
-            method: "DELETE",
-        })
-            .then(res => res.json())
-            .then(data => {
-                const remaining = products.filter((product) => product._id !== id);
-                setProducts(remaining)
+        const confirm = window.confirm('Are You Sure Want To Delete??')
+        if (confirm) {
+            const url = `http://localhost:5000/inventory/${id}`
+            fetch(url, {
+                method: "DELETE",
             })
+                .then(res => res.json())
+                .then(data => {
+                    const remaining = products.filter((product) => product._id !== id);
+                    setProducts(remaining)
+                })
+        }
     }
     return (
         <div>
